feat(metrics): record retries in reducer metrics
Make reducer state the single source of truth for run-level execution metrics and final summary output.

- Add RunMetrics to PipelineState and preserve it across checkpoint serialization
- Increment metrics during reduction for dev/review/fix attempts, commits, retries, and fallbacks
- Derive the final banner summary from reducer metrics to avoid drift from runtime counters
- Add unit/integration coverage for metrics preservation and summary derivation

+++ tests/integration_tests/workflows/continuation_budget.rs
@@ -70,7 +70,7 @@ fn test_fix_continuation_budget_enforced() {
                 }),
             );
             assert_eq!(state.continuation.fix_continuation_attempt, i);
-            assert_eq!(state.metrics.fix_continuations_total, i as u32);
+            assert_eq!(state.metrics.fix_continuations_total, i);
         }
 
         // After fix_continuation_attempt = 2, which >= max_fix_continue_count (2), exhausted
